Fixed OSX porting of dem_interpolate, rtklib, zip lib projects

+++ dem_interpolate/dsm.h
@@ -159,7 +159,7 @@ private:
 	mutable TRIANGLE	_tr;
 	mutable NODE		_nd;
 	SEGMENT				_sg;
-	__int64				_pos;
+    long long				_pos;
 	std::vector<float> quote;
 };
